pyproject.toml
[project]
name = "frynd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.31.0",
    "amadeus>=9.3.0",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.2.3",
    "langchain-community>=0.2.3",
    "langchain-text-splitters>=0.2.0",
    "langchain-google-genai>=1.0.5",
    "google-generativeai>=0.5.3",
    "pydantic>=2.6.0",
    "selenium>=4.38.0",
    "chromadb>=1.3.5",
    "pypdf>=6.4.0",
    "google-cloud-texttospeech>=2.33.0",
    "gtts>=2.5.4",
    "google-genai>=1.54.0",
]
